requirements.txt
flipside==2.0.0
httpx==0.28.1
mcp[cli]>=1.9.0
python-dotenv==1.1.0

